:hammer: refactored components, new files
